+++ SurgSim/Graphics/UnitTests/OsgMeshRepresentationTests.cpp
@@ -36,7 +36,7 @@ using SurgSim::Framework::Runtime;
 namespace
 {
 	std::vector<Vector3d> cubeVertices;
-	std::vector<unsigned int> cubeTriangles;
+	std::vector<size_t> cubeTriangles;
 	std::vector<Vector4d> cubeColors;
 	std::vector<Vector2d> cubeTextures;
 }
